getTag in its own file now
