+++ test/test.js
@@ -12,13 +12,13 @@ import cloneDeepFully from '../src/clone-deep-fully/clone-deep-fully.js';
 import useCustomizers from '../src/use-customizers.js';
 
 import { CLONE, Tag } from '../src/utils/constants.js';
+import { getTag } from '../src/utils/get-tag.js';
 import {
     createFileList,
     getSupportedPrototypes,
     getTypedArrayConstructor
 } from '../src/utils/helpers.js';
 import {
-    getTag,
     isDOMMatrix,
     isDOMMatrixReadOnly,
     isDOMPoint,
